Implement first version of symbolic tracer
The simple script `trace_symbols.py` takes an executable as an argument
and does the following:

 1. Executes the program natively (starting at `main`) and records a
    trace of every instruction executed, stopping when exiting `main`.

 2. Tries to follow this trace of instructions concolically (keeps a
    concrete program state from a native execution in parallel to a
symbolic program state), recording after each instruction the changes it
has made to the program state before that instruction.

 3. Writes the program state at each instruction to log files; writes
    the concrete state of the real execution to 'concrete.log' and the
symbolic difference to 'symbolic.log'.

This first version is a proof of concept and is very fragile. As soon as
angr can't handle a branch instruction (which is the case for almost any
branch instruction), it aborts with an error.