Don't record concrete states up-front
Step the concrete target iteratively along with the symbolic execution.
This allows us to record concrete states when they're needed, and only
one at a time. We still need to record a trace of addresses visited
up-front (at least I think so; look into this).

This is essential for larger binaries, where hundreds of thousands of
concrete states can easily consume all memory on the machine.