Don't pre-record a concrete trace
Do the same thing to the pre-recorded concrete trace that we did to
disassembly and concrete states: Step through it on-the-fly during
execution instead. This saves a lot of time.

Move program counter evaluation from concrete state in `run_block` a bit
to make basic blocks smaller in some cases.